second and probably last update

+++ N3DSReader/Structures/NCSD/PartitionMetadata.cs
@@ -8,16 +8,16 @@ namespace N3DSReader.Structures.NCSD
     {
         public int Id { get; }
         public string Name { get; }
-        public int Offset { get; }
-        public int Length { get; }
+        public long Offset { get; }
+        public long Length { get; }
 
         public PartitionMetadata(int id, byte[] bytes)
         {
             if (bytes.Length != 8) throw new ArgumentException("There should be 8 bytes!");
             Id = id;
             Name = Partitions.Names[Id];
-            Offset = BitConverter.ToInt32(bytes[0..4].ToArray()) * 512;
-            Length = BitConverter.ToInt32(bytes[4..8].ToArray()) * 512;
+            Offset = BitConverter.ToInt32(bytes[0..4].ToArray()) * (long)512;
+            Length = BitConverter.ToInt32(bytes[4..8].ToArray()) * (long)512;
         }
 
         public override string ToString()
